Move the AllMeetingsICalService into its own namespace
