+++ modules/meeting/app/services/all_meetings/ical_service.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# -- copyright
+#-- copyright
 # OpenProject is an open source project management software.
 # Copyright (C) the OpenProject GmbH
 #
@@ -26,24 +26,46 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #
 # See COPYRIGHT and LICENSE files for more details.
-# ++
-module Meetings
-  class ICalController < ApplicationController
-    skip_before_action :check_if_login_required
-    authorization_checked! :index
+#++
+require "icalendar"
+require "icalendar/tzinfo"
+module AllMeetings
+  class ICalService
+    include Meetings::ICalHelpers
+    attr_reader :user, :url_helpers, :include_historic, :calendar, :timezone
 
-    def index
-      token = Token::ICalMeeting.find_by_plaintext_value(params[:token]) # rubocop:disable Rails/DynamicFindBy
+    def initialize(user:, include_historic: false)
+      @user = user
+      @timezone = user.time_zone
+      @url_helpers = OpenProject::StaticRouting::StaticUrlHelpers.new
+      @include_historic = include_historic
+      @calendar = build_icalendar(Time.current.in_time_zone(user.time_zone))
+    end
 
-      user = token.user
-
-      service = AllMeetings::ICalService.new(user:)
-
-      respond_to do |format|
-        format.ics do
-          send_data(service.call.result, filename: "mycal.ics", disposition: "inline; filename=mycal.ics", type: "text/calendar")
-        end
+    def call
+      single_meetings.each do |meeting|
+        build_single_meeting(meeting)
       end
+
+      ServiceResult.success(result: calendar.to_ical)
+    end
+
+    private
+
+    def ical_subject(meeting)
+      "[#{meeting.project.name}] #{I18n.t(:label_meeting)}: #{meeting.title}"
+    end
+
+    def recurring_meetings
+      @recurring_meetings ||= RecurringMeeting.visible(user)
+    end
+
+    def single_meetings
+      @single_meetings ||= if include_historic
+                             Meeting.not_recurring.visible(user)
+                           else
+                             Meeting.not_recurring.from_today.visible(user)
+                           end
     end
   end
 end
